services/collector/requirements.txt
psycopg2-binary
requests
beautifulsoup4
pandas